% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubs.R
\name{hub_table}
\alias{hub_table}
\title{Retrieve a remotely read parquet file of Bioconductor AnnotationHub or
ExperimentHub sqlite table}
\usage{
hub_table(
  hub = c("annotationhub", "experimenthub"),
  tblname = c("resource_full", "resource_metadata", "resource_downloadlinks",
    "input_sources", "location_prefixes", "rdatapaths", "resources", "statuses", "tags")
)
}
\arguments{
\item{hub}{A valid hub name: "annotationhub" or "experimenthub"}

\item{tblname}{A valid parquet file name.}
}
\value{
An Arrow Table containing the requested metadata
}
\description{
Traditionally using the AnnotationHub or ExperimentHub
Bioconductor packages required downloaded an ever-increasing SQLite file 
containing metadata. The SQLite tables have been converted to parquet files
and can be read remotely using the arrow package.
}
\details{
Traditionally using the AnnotationHub or ExperimentHub
Bioconductor packages required downloaded an ever-increasing SQLite file
containting metadata. The SQLite tables have been converted to parquet files
can can be read remotely using the arrow package. The resulting object is
returned as an Arrow Table and is compatible with many dplyr operations.
}
\examples{
statuses <- hub_table("annotationhub", "statuses")
statuses

}
\author{
Lori Shepherd
}
